+++ app/contact/page.tsx
@@ -5,8 +5,8 @@ import { siX } from "simple-icons";
 
 import { SimpleIcon } from "../components/ui/icons/SimpleIcon";
 import { ogImageUrl, WithSiteTitle } from "../constants";
-import { Footer } from "../features/layout/Footer";
-import { Header } from "../features/layout/Header";
+import { Footer } from "../features/layout/Footer/Footer";
+import { Header } from "../features/layout/Header/Header";
 import hoverStyles from "../styles/card-hover.module.css";
 
 import styles from "./page.module.css";
