--- app/features/analytics/GoogleAnalytics/index.ts
@@ -1,1 +1,0 @@
-export { GoogleAnalytics } from "./GoogleAnalytics";
