--- app/features/analytics/WebVitals/index.ts
@@ -1,1 +1,0 @@
-export { WebVitals } from "./WebVitals";
