--- app/features/posts/ArticleGrid/index.ts
@@ -1,1 +1,0 @@
-export { ArticleGrid } from "./ArticleGrid";
