--- app/features/theme/ThemeProvider/index.ts
@@ -1,1 +1,0 @@
-export { ThemeProvider, useTheme } from "./ThemeProvider";
